Renaming Paragraph to Text and updating docs and references
