+++ src/components/text/types.ts
@@ -4,9 +4,12 @@
  * "An interface can only extend an identifier/qualified-name with optional type
  * arguments. ts(2499)"
  */
-type ParagraphElementProps = JSX.IntrinsicElements['p']
+type TextElementProps =
+  | JSX.IntrinsicElements['p']
+  | JSX.IntrinsicElements['span']
 
-export interface ParagraphProps extends ParagraphElementProps {
+export type TextProps = TextElementProps & {
+  asElement: 'p' | 'span'
   size?: 100 | 200 | 300
   weight?: 'regular' | 'medium' | 'semibold' | 'bold'
 }
